Bullets!
git-svn-id: svn+ssh://leetcode.net/home/jreese/backup/srv/svn/repos/rit/2dgames/trunk@521 ad4848e8-a12d-0410-9a8f-f544a64f585b

+++ SnailsPace/SnailsPace/Objects/GameObject.cs
@@ -14,7 +14,7 @@ namespace SnailsPace.Objects
 		public float maxVelocity;
        
         // The sprite's layer and parallax.
-        public int layer;       // 0: background ... 5: foreground
+        public float layer;       // 0: background ... 5: foreground
         
         // Sprites that make up the game object.
         public Dictionary<String, Sprite> sprites;
